refactor: move down the inline script searching logic
Because the RegExp execution could be very expensive.
So when the current script is not inline but somehow it falls into this
code branch too early (could be an unintended bug or a rare edge case),
the RegExp won't block the page execution, and the function would return
early at the `readyState` branch.